requirements.txt
CelestePythonInterface @ git+https://github.com/hdrien0/Celeste-Instrumentation.git#subdirectory=CelestePythonInterface
contourpy==1.2.0
cycler==0.12.1
dict2xml==1.7.4
dill==0.3.8
fonttools==4.47.2
graphviz==0.20.1
kiwisolver==1.4.5
matplotlib==3.8.2
neat-python==0.92
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0